requirements.txt
asgiref==3.7.2
Django==4.2.6
django-phonenumber-field==7.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
inflection==0.5.1
packaging==23.2
phonenumbers==8.13.24
PyJWT==2.8.0
pytz==2023.3.post1
PyYAML==6.0.1
setuptools==68.2.2
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
